+++ TablePro/Views/Main/Extensions/MainContentView+EventHandlers.swift
@@ -35,7 +35,7 @@ extension MainContentView {
         guard !coordinator.isTearingDown else { return }
         let aggregated = MainContentCoordinator.aggregatedTabs(for: coordinator.connectionId)
         coordinator.persistence.saveNow(
-            tabs: aggregated,
+            windowedTabs: aggregated,
             selectedTabId: newTabId
         )
         MainContentView.lifecycleLogger.debug(